requirements.txt
pytest
pylint
pyyaml
anthropic
python-dotenv